+++ CHANGELOG.md
@@ -22,7 +22,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Removed
 - Drop `AccessDeniedException` (#759)
-- Drop the legacy configuration check (#758)
+- Drop the legacy configuration check (#758, #760)
 - Drop deprecated methods from `TemplateHelper` (#741, #742)
 - Drop the use of removed Core functionality (#735, #736)
 - Drop support for `$GLOBALS['TYPO3_CONF_VARS']['EXT']['extConf']` (#728)